python/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.1
blinker==1.6.2
celery==5.4.0
certifi==2023.7.22
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cron-descriptor==1.4.5
cycler==0.12.0
Django==4.2.20
django-appconf==1.0.6
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.11.2
django-filter==24.3
django-geojson==4.1.0
django-livereload-server==0.5.1
django-select2==8.2.3
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
Fiona==1.9.4.post1
Flask==2.3.3
fonttools==4.43.0
GDAL==3.9.0
geojson==3.1.0
geopandas==0.14.0
geoserver-rest==2.5.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
Js2Py==0.74
kiwisolver==1.4.5
kombu==5.5.4
livereload==2.7.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
packaging==25.0
pandas==2.1.1
Pillow==10.0.1
pipwin==0.5.2
prompt_toolkit==3.0.51
psycopg2==2.9.7
Pygments==2.16.1
pyjsparser==2.7.1
PyJWT==2.9.0
pymongo==4.0
PyOpenGL==3.1.7
pyparsing==3.1.1
PyPrind==2.11.3
pyproj==3.6.1
pySmartDL==1.3.4
python-crontab==3.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==6.2.0
requests==2.31.0
seaborn==0.13.0
shapely==2.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.3.1
tornado==6.4.1
typing_extensions==4.8.0
tzdata==2025.2
tzlocal==5.0.1
urllib3==2.0.5
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.7
xmltodict==0.14.2
